+++ src/index.js
@@ -1013,6 +1013,18 @@ Object.defineProperty(exports, "SpeechBubbleIcon", {
     return _speechBubbleIcon.SpeechBubbleIcon;
   }
 });
+Object.defineProperty(exports, "AirplaneIconNext", {
+  enumerable: true,
+  get: function get() {
+    return _airplaneIconNext.AirplaneIconNext;
+  }
+});
+Object.defineProperty(exports, "BubbleIconNext", {
+  enumerable: true,
+  get: function get() {
+    return _bubbleIconNext.BubbleIconNext;
+  }
+});
 Object.defineProperty(exports, "CheckedShieldIconNext", {
   enumerable: true,
   get: function get() {
@@ -1559,6 +1571,12 @@ Object.defineProperty(exports, "TagButton", {
     return _tagButton.TagButton;
   }
 });
+Object.defineProperty(exports, "TextButton", {
+  enumerable: true,
+  get: function get() {
+    return _textButton.TextButton;
+  }
+});
 Object.defineProperty(exports, "buttonModifierStyles", {
   enumerable: true,
   get: function get() {
@@ -1581,6 +1599,12 @@ Object.defineProperty(exports, "ClickableCard", {
   enumerable: true,
   get: function get() {
     return _clickableCard.ClickableCard;
+  }
+});
+Object.defineProperty(exports, "ContentCard", {
+  enumerable: true,
+  get: function get() {
+    return _contentCard.ContentCard;
   }
 });
 Object.defineProperty(exports, "ContributionCard", {
@@ -2606,6 +2630,10 @@ var _statsIcon = require("./components/graphics/icons/stats-icon");
 
 var _speechBubbleIcon = require("./components/graphics/icons/speech-bubble-icon");
 
+var _airplaneIconNext = require("./components/graphics/icons-next/airplane-icon-next");
+
+var _bubbleIconNext = require("./components/graphics/icons-next/bubble-icon-next");
+
 var _checkedShieldIconNext = require("./components/graphics/icons-next/checked-shield-icon-next");
 
 var _chronoIconNext = require("./components/graphics/icons-next/chrono-icon-next");
@@ -2764,6 +2792,8 @@ var _socialButtonIcon = require("./components/molecules/buttons/social-button-ic
 
 var _tagButton = require("./components/molecules/buttons/tag-button");
 
+var _textButton = require("./components/molecules/buttons/text-button");
+
 var _modifierStyles2 = require("./components/molecules/buttons/button/helpers/modifier-styles");
 
 var _backerCard = require("./components/molecules/cards/backer-card");
@@ -2771,6 +2801,8 @@ var _backerCard = require("./components/molecules/cards/backer-card");
 var _backingCard = require("./components/molecules/cards/backing-card");
 
 var _clickableCard = require("./components/molecules/cards/clickable-card");
+
+var _contentCard = require("./components/molecules/cards/content-card");
 
 var _contributionCard = require("./components/molecules/cards/contribution-card");