+++ CHANGELOG.md
@@ -5,6 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Logo: Add `YouMatterLogo` logo.
+Features: 
+  - `RadioButton`: Add `small` to the size prop.
+  - `RadioButtonSet`: Add `small` to the size prop.
+  - `ContentCard`: Add component.
+  - `TextButton`: Add component.
+  - `AirplaneIconNext`: Add component.
+  - `BubbleIconNext`: Add component.
 
 ## [6.6.0] - 2021-12-02